requirements.txt
flask==3.0.3
pydantic==2.7.1
pymongo==4.7.0
pydantic-settings==2.2.1
bunnet==1.3.0
pyproj==3.6.1
gunicorn==22.0.0
celery==5.4.0
httpx==0.27.0
pytest_httpx==0.30.0
